+++ src/pages/Auth/SignUpPageSocial/hooks/useSignUpPageSocial.ts
@@ -9,7 +9,7 @@ export function useSignUpPageSocial() {
 		async (_provider: string) => {
 			setLoading(true);
 			try {
-				navigate("/main");
+				navigate("/index");
 			} catch (error) {
 				console.error("로그인 오류:", error);
 				alert("로그인에 실패했습니다.");
